wrs: Introduce `IsNewMessage()` function
A new function to utilize the `MSG_MEMORY_DIR` config.

It's a simple file-based memory to keep track of sent messages with the
help of filesystem. Hash the message, check if the file exists in the
memory directory, if it exists, skip the message. If it does not exist,
create the file to remember the message.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>